DBF

This function is an internal function used by DBFMCL to detect informative elements (i.e., those that belong to dense regions). User should not use this function. Instead they can use the DBFMCL function with clustering argument set to FALSE.

See DBFMCL

DBF(
  data,
  output_path = ".",
  name = NULL,
  distance_method = "pearson",
  silent = FALSE,
  k = 100,
  fdr = 10,
  seed = 123
)

Arguments

data

a matrix or data.frame

output_path

a character string representing the data directory where output files will be stored. Default to current working directory.

name

a prefix for the file name

distance_method

a method to compute the distance to the k-th nearest neighbor.

silent

if set to TRUE (default), the progression of distance matrix calculation is not displayed.

k

the neighborhood size.

fdr

a value for the false discovery rate.

seed

specify seeds for random number generator.

Warnings

Works only on UNIX-alikes platforms.

References

Lopez F.,Textoris J., Bergon A., Didier G., Remy E., Granjeaud S., Imbert J. , Nguyen C. and Puthier D. TranscriptomeBrowser: a powerful and flexible toolbox to explore productively the transcriptional landscape of the Gene Expression Omnibus database. PLoSONE, 2008;3(12):e4001.

See also

Author

Bergon A., Bavais J., Textoris J., Granjeaud S., Lopez F and Puthier D.