The "Density Based Filtering and Markov CLustering" algorithm (DBF-MCL).

DBFMCL is a tree-steps adaptative algorithm that (i) find elements located in dense areas (DBF), (ii) uses selected items to construct a graph, (iii) performs graph partitioning using the Markov CLustering Algorithm (MCL).

This function requires installation of the mcl program (http://www.micans.org/mcl). See "Warnings" section for more informations.

When analyzing a noisy dataset, one is interested in isolating dense regions as they are populated with genes/elements that display weak distances to their nearest neighbors (i.e. strong profile similarities). To isolate these regions DBF-MCL computes, for each gene/element, the distance with its kth nearest neighbor (DKNN).In order to define a critical DKNN value that will depend on the dataset and below which a gene/element will be considered as falling in a dense area, DBF-MCL computes simulated DKNN values by using an empirical randomization procedure. Given a dataset containing n genes and p samples, a simulated DKNN value is obtained by sampling n distance values from the gene-gene distance matrix D and by extracting the kth-smallest value. This procedure is repeated n times to obtain a set of simulated DKNN values S. Computed distributions of simulated DKNN are used to compute a FDR value for each observed DKNN value. The critical value of DKNN is the one for which a user-defined FDR value (typically 10%) is observed. Genes with DKNN value below this threshold are selected and used to construct a graph. In this graph, edges are constructed between two genes (nodes) if one of them belongs to the k-nearest neighbors of the other. Edges are weighted based on the respective coefficient of correlation (i.e., similarity) and the graph obtained is partitioned using the Markov CLustering Algorithm (MCL).

DBFMCL(
  data = NULL,
  filename = NULL,
  path = ".",
  output_path = ".",
  mcl_threads = 1,
  name = NULL,
  distance_method = "pearson",
  av_dot_prod_min = 2,
  min_cluster_size = 10,
  silent = FALSE,
  k = 50,
  fdr = 10,
  inflation = 8,
  seed = 123
)

Arguments

data

a matrix, data.frame or Seurat object.

filename

a character string representing the file name.

path

a character string representing the data directory where intermediary files are to be stored. Default to current working directory.

output_path

a character string representing the data directory where output files will be stored. Default to current working directory.

mcl_threads

An integer to determine number of threads for mcl algorithm.

name

a prefix for the names of the intermediary files created by DBF and MCL.

distance_method

a method to compute the distance to the k-th nearest neighbor.

av_dot_prod_min

Any cluster with average dot product below this value is discarded. This allow to delete clusters in which correlation is influenced/supported by very few samples (typically 1).

min_cluster_size

Minimum number of element inside a cluster. MCL tend to create lots of clusters with very few (e.g 2) objects.

silent

if set to TRUE, the progression of distance matrix calculation is not displayed.

k

the neighborhood size.

fdr

an integer value corresponding to the false discovery rate (range: 0 to 100).

inflation

the main control of MCL. Inflation affects cluster granularity. It is usually chosen somewhere in the range [1.2-5.0]. inflation = 5.0 will tend to result in fine-grained clusterings, and whereas inflation = 1.2 will tend to result in very coarse grained clusterings. By default, inflation = 2.0. Default setting gives very good results for microarray data when k is set between 70 and 250.

seed

specify seeds for random number generator.

Value

a ClusterSets class object.

Warnings

With the current implementation, this function only works only on UNIX-like plateforms.

MCL should be installed. One can used the following command lines in a terminal:

# Download the latest version of mcl (the script has been tested successfully with the 06-058 version). wget http://micans.org/mcl/src/mcl-latest.tar.gz # Uncompress and install mcl tar xvfz mcl-latest.tar.gz cd mcl-xx-xxx ./configure make sudo make install # You should get mcl in your path mcl -h

References

- Van Dongen S. (2000) A cluster algorithm for graphs. National Research Institute for Mathematics and Computer Science in the 1386-3681. - Lopez F.,Textoris J., Bergon A., Didier G., Remy E., Granjeaud S., Imbert J. , Nguyen C. and Puthier D. TranscriptomeBrowser: a powerful and flexible toolbox to explore productively the transcriptional landscape of the Gene Expression Omnibus database. PLoSONE, 2008;3(12):e4001.

Author

Bergon A., Bavais J., Textoris J., Granjeaud S., Lopez F and Puthier D.

Examples

if (FALSE) {
  m <- matrix(rnorm(80000), nc=20)
  m[1:100,1:10] <- m[1:100,1:10] + 4
   m[101:200,11:20] <- m[101:200,11:20] + 3
   m[201:300,5:15] <- m[201:300,5:15] + -2
   res <- DBFMCL(data=m,
                 distance_method="pearson",
                 av_dot_prod_min = 0,
                 inflation = 1.2,
                 k=25,
                 fdr = 10)
plot_clust(res, ceil = 10, floor = -10)
plot_clust(res, type="tile", ceil = 10, floor = -10)
write_clust(res, filename_out = "ALL.sign.txt")
}