Perform Community Detection Using the infomap Algorithm (internal function)

Applies the infomap algorithm for community detection on a graph derived from the input data in a `ClusterSet` object. Outputs the resulting clusters to a specified file and updates the parameters in the `ClusterSet` object.

call_infomap_clusterset(object, modularity = FALSE, nb.trials = 10)

Arguments

object

A `ClusterSet` object.

modularity

See igraph::cluster_infomap().

nb.trials

See igraph::cluster_infomap().

Value

The modified `ClusterSet` object with updated parameters, including the infomap algorithm results

Details

This function reads a graph from the input file defined in the `ClusterSet` object, processes the graph using the infomap community detection algorithm from the `igraph` package, and return the clusters enclose in the `ClusterSet`.

Examples

# Restrict vebosity to info messages only.
set_verbosity(1)
# Load a dataset
load_example_dataset("7871581/files/pbmc3k_medium")
#> |-- INFO :  Dataset 7871581/files/pbmc3k_medium was already loaded. 
# Select informative genes
res <- select_genes(pbmc3k_medium,
                    distance = "pearson",
                    row_sum=5)
#> |-- INFO :  Computing distances using selected method: pearson 
#> |-- INFO :  Computing distances to KNN. 
#> |-- INFO :  Computing simulated distances to KNN. 
#> |-- INFO :  Computing threshold of distances to KNN (DKNN threshold). 
#> |-- INFO :  Selecting informative genes. 
#> |-- INFO :  Creating the ClusterSet object. 

# Cluster informative features
res <- gene_clustering(res, method="closest_neighborhood",
                       algorithm="infomap")
#> |-- INFO :  Computing distances between selected genes. 
#> |-- INFO :  Compute distances to KNN between selected genes. 
#> |-- INFO :  Computing distances to KNN. 
#> |-- INFO :  Creating the input file for graph partitioning. 
#> |-- INFO :  Deleting reciprocal edges. 
#> |-- INFO :  Converting distances into weights. 
#> |-- INFO :  Writing the input file. 
#> |-- INFO :  Input file saved in :/var/folders/zy/wl3dj2_n76zfc8sdvny1q06c0000gn/T//Rtmp12r3vc/R14z6ff8TV.graph_input.txt 
#> |-- INFO :  The infomap algorithm has been selected. 
#> |-- INFO :  Calling infomap algorithm.