Generate a discrete color palette

This function generates a vector of colors for a discrete variable, given a specified palette name.

discrete_palette(n = 10, palette = c("Ju1", "De1", "ggplot"))

Arguments

n

An integer specifying the number of colors to generate.

palette

A character vector specifying the palette to use.

Value

A character vector of color codes.

Examples

discrete_palette()
#>         1         2         3         4         5         6         7         8 
#> "#9F1717" "#517416" "#9D1C70" "#FFCA3A" "#9579B9" "#DA7316" "#1882C0" "#EF9292" 
#>         9        10 
#> "#B6E36A" "#EA8AC9" 
discrete_palette(n=20, palette = "ggplot")
#>         1         2         3         4         5         6         7         8 
#> "#F8766D" "#EA8331" "#D89000" "#C09B00" "#A3A500" "#7CAE00" "#39B600" "#00BB4E" 
#>         9        10        11        12        13        14        15        16 
#> "#00BF7D" "#00C1A3" "#00BFC4" "#00BAE0" "#00B0F6" "#35A2FF" "#9590FF" "#C77CFF" 
#>        17        18        19        20 
#> "#E76BF3" "#FA62DB" "#FF62BC" "#FF6A98"