Enrichment analysis using GO database of gene clusters from a ClusterSet object

This function performs enrichment analysis of all gene clusters from a ClusterSet object and store the results in the cluster_annotations slot.

# S4 method for ClusterSet
enrich_go(object, species = "Hsapiens", ontology = "all")

Arguments

object

an object of class ClusterSet.

species

a character string indicating the species name, as a concatenation of the first letter of the name (uppercase) and the family name (lowercase), e.g human -> Hsapiens, mouse -> Mmusculus

ontology

a character string indicating the ontology to use for the enrichment analysis. One of "BP", "MF", and "CC" ontology, or "ALL".

Value

A ClusterSet object

Examples

load_example_dataset("7871581/files/pbmc3k_medium_clusters")
#> |-- INFO :  Dataset 7871581/files/pbmc3k_medium_clusters was already loaded. 
load_example_dataset("8028126/files/pbmc3k_medium_clusters_enr")
#> |-- INFO :  Dataset 8028126/files/pbmc3k_medium_clusters_enr has been loaded. 
# pbmc3k_medium_clusters_enr <- enrich_go(pbmc3k_medium_clusters)
pbmc3k_medium_clusters_enr
#> 		An object of class ClusterSet
#> 		Name: R14z6ff8TV 
#> 		Memory used:  1240735344 
#> 		Number of cells:  361 
#> 		Number of informative genes:  291 
#> 		Number of gene clusters:  15 
#> 		This object contains the following informations:
#> 			 -  data 
#> 			 -  gene_clusters 
#> 			 -  top_genes 
#> 			 -  gene_clusters_metadata 
#> 			 -  gene_cluster_annotations 
#> 			 -  cells_metadata 
#> 			 -  dbf_output 
#> 			 -  parameters 
#> 				 *  distance_method  =  pearson 
#> 				 *  k  =  80 
#> 				 *  highest  =  5e-05 
#> 				 *  fdr  =  0.005 
#> 				 *  row_sum  =  5 
#> 				 *  no_dknn_filter  =  FALSE 
#> 				 *  seed  =  123 
#> 				 *  keep_nn  =  FALSE 
#> 				 *  k_mcl_graph  =  5 
#> 				 *  output_path  =  /var/folders/zy/wl3dj2_n76zfc8sdvny1q06c0000gn/T//RtmpuPDGyK 
#> 				 *  name  =  R14z6ff8TV 
#> 				 *  inflation  =  2.2 
is(pbmc3k_medium_clusters_enr@gene_cluster_annotations)
#> [1] "list"                                         
#> [2] "vector"                                       
#> [3] "OptionalList"                                 
#> [4] "AssayData"                                    
#> [5] "list_OR_List"                                 
#> [6] "character or array or list or NULL or logical"
#> [7] "vector_OR_Vector"                             
#> [8] "vector_OR_factor"