Filter cluster from a ClusterSet object using dot product.

This function filters clusters of gene expression data based on their dot products. It aims to remove clusters that have a lot of zeros and are supported by only a few cells or spots. To do this, the function first converts the gene expression data for each cluster into a binary form (values greater than 1 are set to 1). Then it calculates the dot product for this binary matrix, which produces a gene-gene matrix showing the number of cells/spots where each pair of genes are expressed together. The function then calculates the median value of the maximum concordances across all genes, which can be used to determine whether a cluster should be filtered out or not.

filter_by_dot_prod(object = NULL, av_dot_prod_min = 2)

Arguments

object

A ClusterSet object.

av_dot_prod_min

Any cluster with average dot product below this value is discarded. This allow to delete clusters in which correlation is influenced/supported by very few samples (typically 1).

Examples

load_example_dataset("7871581/files/pbmc3k_medium_clusters")
#> |-- INFO :  Dataset 7871581/files/pbmc3k_medium_clusters was already loaded. 
pbmc3k_medium_clusters <- top_genes(pbmc3k_medium_clusters)
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 6. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 7. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 8. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 9. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 10. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 11. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 12. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 13. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 14. All genes will be used. 
#> |-- INFO :  Number of top genes is greater than the number of genes in cluster 15. All genes will be used. 
#> |-- INFO :  Results are stored in top_genes slot of ClusterSet object. 
nclust(pbmc3k_medium_clusters)
#> [1] 15
obj <- filter_by_dot_prod(pbmc3k_medium_clusters, av_dot_prod_min=5)
nclust(obj) 
#> [1] 9