Filter out cluster supported by few cells.

This function filters out the gene clusters based on the number of cells expressing a certain percentage of genes present in this gene cluster.

filter_nb_supporting_cells(
  object = NULL,
  min_nb_supporting_cell = 3,
  min_pct_gene_expressed = 50
)

Arguments

object

A ClusterSet object.

min_nb_supporting_cell

An integer indicating the minimum number of cell supporting a cluster. A cell supports a cluster if it expresses at least min_pct_gene_expressed % of the genes from the cluster.

min_pct_gene_expressed

See min_nb_supporting_cell argument.

Value

A ClusterSet object where clusters that did not pass the filter have been removed.

Examples

load_example_dataset("7871581/files/pbmc3k_medium_clusters")
#> |-- INFO :  Dataset 7871581/files/pbmc3k_medium_clusters was already loaded. 
new_cs <- filter_nb_supporting_cells(pbmc3k_medium_clusters, 3, 50)
#> |-- INFO :  2 clusters with less than 3 cells expressing at least 50% of genes were filtered out. 
#> |-- INFO :  Number of clusters left 13