Construct a graph based on selected genes in a ClusterSet object and their neighbors

This function creates a graph based on the gene selected by select_genes() and their neighbors. It generates an input file required for the MCL (Markov Cluster) algorithm.

keep_dbf_graph(object = NULL, output_path = NULL, name = NULL)

Arguments

object

A ClusterSet object.

output_path

a character indicating the path where the output files will be stored.

name

a character string giving the name for the output files. If NULL, a random name is generated.

Value

The output of this function is a tab-separated file, which serves as input for the MCL algorithm.

Examples

# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Create a matrix with 4 signatures
m <- create_4_rnd_clust()

# Select informative features
res <- select_genes(m[1:600,],
                    distance = "pearson",
                    k = 75,
                    noise_level = 0.3,
                    fdr = 1e-8,
                    row_sum = -Inf)
#> |-- INFO :  Computing distances using selected method: pearson 
#> |-- INFO :  Computing distances to KNN. 
#> |-- INFO :  Computing simulated distances to KNN. 
#> |-- INFO :  Computing threshold of distances to KNN (DKNN threshold). 
#> |-- INFO :  Selecting informative genes. 
#> |-- INFO :  Creating the ClusterSet object. 
                    
# Construct a graph based on genes selected with select_genes() and their neighbors
keep_dbf_graph(object = res)
#> |-- INFO :  Creating the input file for MCL algorithm. 
#> |-- INFO :  Writing the input file. 
#> |-- INFO :  The input file saved in '/var/folders/zy/wl3dj2_n76zfc8sdvny1q06c0000gn/T/RtmpIMWbYW/06oFCM0di0.input_mcl.txt'. 
#> 		An object of class ClusterSet
#> 		Name: 06oFCM0di0 
#> 		Memory used:  1743336 
#> 		Number of cells:  20 
#> 		Number of informative genes:  259 
#> 		Number of gene clusters:  1 
#> 		This object contains the following informations:
#> 			 -  data 
#> 			 -  gene_clusters 
#> 			 -  top_genes 
#> 			 -  gene_clusters_metadata 
#> 			 -  gene_cluster_annotations 
#> 			 -  cells_metadata 
#> 			 -  dbf_output 
#> 			 -  parameters 
#> 				 *  distance_method  =  pearson 
#> 				 *  k  =  75 
#> 				 *  noise_level  =  0.3 
#> 				 *  fdr  =  1e-08 
#> 				 *  row_sum  =  -Inf 
#> 				 *  no_dknn_filter  =  FALSE 
#> 				 *  seed  =  123 
#> 				 *  keep_nn  =  TRUE 
#> 				 *  output_path  =  /var/folders/zy/wl3dj2_n76zfc8sdvny1q06c0000gn/T//RtmpIMWbYW 
#> 				 *  name  =  06oFCM0di0