Invokes the command line version of Markov CLustering algorithm (MCL).

This function invokes the mcl system command. MCL is a clustering algorithm for graphs that was developped by Stijn van Dongen (see references for further informations).

mcl_system_cmd(
  name,
  inflation = 2,
  input_path = ".",
  silent = FALSE,
  threads = 1
)

Arguments

name

a character string corresponding to the file name.

inflation

the main control of MCL. Inflation affects cluster granularity. It is usually chosen somewhere in the range [1.2-5.0]. inflation = 5.0 will tend to result in fine-grained clusterings, and whereas inflation = 1.2 will tend to result in very coarse grained clusterings. By default, inflation = 2.0. Default setting gives very good results for microarray data when k is set around 100.

input_path

a character string representing the directory path of the input file used by mcl. Default is the current working directory.

silent

if set to TRUE, the progression of the MCL partitionning is not displayed.

threads

The number of threads to use.

Value

Returns a file with the ".mcl\_out.txt" extension.

warning

Works only on UNIX-like plateforms. MCL should be installed. The following command lines can be used for installation.

# Download the latest version of mcl (RTools4TB has been tested successfully with the 06-058 version). wget http://micans.org/mcl/src/mcl-latest.tar.gz # Uncompress and install mcl tar xvfz mcl-latest.tar.gz cd mcl-xx-xxx ./configure make sudo make install # You should get mcl in your path mcl -h

References

Stijn van Dongen. A cluster algorithm for graphs. Technical Report INS-R0010, National Research Institute for Mathematics and Computer Science in the Netherlands, Amsterdam, May 2000. http://www.cwi.nl/ftp/CWIreports/INS/INS-R0010.ps.Z

Author

Bergon A., Lopez F., Textoris J., Granjeaud S. and Puthier D.