plot_heatmap

Plot the observed and simulated distance with the Kth nearest neighbors.

plot_heatmap(
  object,
  center = TRUE,
  ceil = 1,
  floor = -1,
  cell_order = NULL,
  cell_ordering_method = "hclust",
  cluster = NULL,
  use_top_genes = FALSE,
  name = NULL,
  xlab = NULL,
  ylab = NULL,
  colorbar_name = "Expression level",
  show_legend = TRUE,
  colors = c("#A9D6E5", "#2166AC", "#000000", "#B2182B", "#FFCA3A"),
  row_labels = TRUE,
  col_labels = FALSE,
  label_size = 9,
  line_size = 15
)

Arguments

object

A ClusterSet object.

center

A logical to indicate whether to center row..

ceil

A value for ceiling (NULL for no ceiling). Ceiling is performed centering.

floor

A value for flooring (NULL for no flooring). Flooring is performed after centering.

cell_order

A vector of cell names already ordered.

cell_ordering_method

The clustering method to be used. This must be "hclust", ADD OTHER CLUSTERING METHOD.

cluster

A cluster id to plot. Default is NULL for plotting all cluster.

use_top_genes

A logical to indicate whether to use highly similar genes in the slot top_genes of ClusterSet.

name

A title for the heatmap.

xlab

A title for the x axis.

ylab

A title for the y axis.

colorbar_name

A title for the colorbar.

show_legend

A logical to indicate whether to show colorbar.

colors

A vector of colors.

row_labels

A logical to indicate whether to show row labels.

col_labels

A logical to indicate whether to show col labels.

label_size

A value for label font size.

line_size

An integer for the horizontal white line size.

Value

Iheatmap-class object

Examples

m <- matrix(rnorm(40000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- DBFMCL(data=m,
distance_method="pearson",
av_dot_prod_min = 0,
inflation = 2,
k=25,
fdr = 10)
#> |--  Row names not provided. Adding. 
#> |--  Colum names not provided. Adding. 
#> The following parameters will be used : 
#> 	Working directory:  /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference 
#> 	Ouput directory:  /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference 
#> 	Name:  8Z2EHzrk77 
#> 	Distance method:  pearson 
#> 	Minimum average dot product for clusters:  0 
#> 	Minimum cluster size:  10 
#> 	Number of neighbors:  25 
#> 	FDR:  10 % 
#> 	Inflation: 2 
#> 	Visualize standard outputs from both mcl and cluster commands:  FALSE 
#> 
#> |--  DBF completed. Starting MCL step. 
#> Running mcl (graph partitioning)... 
#> |--  Done 
#> |--  creating file : /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference/8Z2EHzrk77.mcl_out.txt 
#> |--  Reading MCL output: /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference/8Z2EHzrk77.mcl_out.txt 
#> |--  3  clusters conserved after MCL partitioning. 
#> |--  0  clusters filtered out from MCL partitioning (size and mean dot product). 

plot_heatmap(object = res)
#> |--  Centering matrix. 
#> |--  Ordering cells based on hierarchical clustering. 
#> |--  Ceiling matrix. 
#> |--  Flooring matrix. 
#> |--  Plotting heatmap. 
plot_heatmap(object = res, cluster = "1")
#> |--  Centering matrix. 
#> |--  Ordering cells based on hierarchical clustering. 
#> |--  Ceiling matrix. 
#> |--  Flooring matrix. 
#> |--  Plotting heatmap.