Select top_genes based on go terms

The clusterSet object contains a top_genes slot that can be used to display genes in heatmaps (see plot_heatmap function). Here the function select top_genes based on GO terms.

# S4 method for class 'ClusterSet'
top_by_go(
  object,
  go_id = "GO:0003677",
  species = "hsapiens",
  gene_id = "hgnc_symbol",
  host = "https://www.ensembl.org",
  as_list = FALSE
)

Arguments

object

A ClusterSet object.

go_id

A GO ID.

species

The target species according to biomaRt::listDatasets(mart = ensembl)$dataset (e.g hsapiens, mmusculus, rnorvegicus...).

gene_id

The type of gene identifier in the clusterSet object (default hgnc_symbol). use biomaRt::listFilters() to get the vailable list.

host

Host to connect to. Defaults to www.ensembl.org.

as_list

Return a list not a ClusterSet object.

Value

A ClusterSet object or a list (see as_list).

Examples

# Set verbosity to 1 to display info messages only.
set_verbosity(1)

# Load a dataset
load_example_dataset('7871581/files/pbmc3k_medium_clusters')
#> |-- INFO :  Dataset 7871581/files/pbmc3k_medium_clusters was already loaded. 

# DNA Binding: "GO:0003677"
#' # DNA Binding: "GO:0003677"
pbmc3k_medium_clusters <- top_by_go(pbmc3k_medium_clusters[1:2, ], go_id = "GO:0003677")
#> Warning: Invalid mirror. Select a mirror from [www, useast, asia].
#> Default when no mirror is specified is to use www.ensembl.org which may be automatically redirected.
#> |-- INFO :  Connecting to Biomart... 
#> |-- INFO :  Processing cluster 1 
#> |-- INFO :  Processing cluster 2 
pbmc3k_medium_clusters@top_genes
#> [[1]]
#> [1] "NACA" "RPL6" "RPS3"
#> 
#> [[2]]
#> [1] "APP"
#>