Best co-expressed genes from each gene cluster

Extract the highly co-expressed genes of each gene cluster.

top_genes(object, top = 20, cluster = "all")

Arguments

object

A ClusterSet object.

top

A value for the number of the most similar genes in gene clusters.

cluster

A vector of gene cluster identity.

Value

Examples

m <- matrix(rnorm(80000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- DBFMCL(data=m,
              distance_method="pearson",
              av_dot_prod_min = 0,
              inflation = 1.2,
              k=25,
              fdr = 10)
#> |--  Row names not provided. Adding. 
#> |--  Colum names not provided. Adding. 
#> The following parameters will be used : 
#> 	Working directory:  /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference 
#> 	Ouput directory:  /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference 
#> 	Name:  X1tR0oi30C 
#> 	Distance method:  pearson 
#> 	Minimum average dot product for clusters:  0 
#> 	Minimum cluster size:  10 
#> 	Number of neighbors:  25 
#> 	FDR:  10 % 
#> 	Inflation: 1.2 
#> 	Visualize standard outputs from both mcl and cluster commands:  FALSE 
#> 
#> |--  DBF completed. Starting MCL step. 
#> Running mcl (graph partitioning)... 
#> |--  Done 
#> |--  creating file : /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference/X1tR0oi30C.mcl_out.txt 
#> |--  Reading and filtering MCL output: /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference/X1tR0oi30C.mcl_out.txt 
#> |--  3  clusters conserved after MCL partitioning. 
#> |--  0  clusters filtered out from MCL partitioning (size and mean dot product). 
              
res <- top_genes(object = res, top=500, cluster="all")
#> Warning: Number of top genes is greater than the number of genes in cluster 1. All genes will be use and order by similarity rank.
#> Warning: Number of top genes is greater than the number of genes in cluster 2. All genes will be use and order by similarity rank.
#> Warning: Number of top genes is greater than the number of genes in cluster 3. All genes will be use and order by similarity rank.
#> |--  Results are stored in top_genes slot of ClusterSet object. 
res@top_genes
#>           gene.top.1 gene.top.2 gene.top.3 gene.top.4 gene.top.5 gene.top.6
#> cluster.1 "gene68"   "gene58"   "gene92"   "gene8"    "gene27"   "gene36"  
#> cluster.2 "gene164"  "gene152"  "gene192"  "gene180"  "gene199"  "gene125" 
#> cluster.3 "gene250"  "gene227"  "gene257"  "gene212"  "gene253"  "gene221" 
#>           gene.top.7 gene.top.8 gene.top.9 gene.top.10 gene.top.11 gene.top.12
#> cluster.1 "gene49"   "gene54"   "gene3"    "gene59"    "gene77"    "gene33"   
#> cluster.2 "gene161"  "gene148"  "gene177"  "gene115"   "gene102"   "gene107"  
#> cluster.3 "gene217"  "gene271"  "gene244"  "gene246"   "gene258"   "gene216"  
#>           gene.top.13 gene.top.14 gene.top.15 gene.top.16 gene.top.17
#> cluster.1 "gene98"    "gene31"    "gene25"    "gene51"    "gene29"   
#> cluster.2 "gene173"   "gene183"   "gene123"   "gene108"   "gene198"  
#> cluster.3 "gene211"   "gene228"   "gene263"   "gene229"   "gene240"  
#>           gene.top.18 gene.top.19 gene.top.20 gene.top.21 gene.top.22
#> cluster.1 "gene15"    "gene87"    "gene74"    "gene28"    "gene12"   
#> cluster.2 "gene154"   "gene179"   "gene191"   "gene140"   "gene168"  
#> cluster.3 "gene267"   "gene265"   "gene279"   "gene204"   "gene280"  
#>           gene.top.23 gene.top.24 gene.top.25 gene.top.26 gene.top.27
#> cluster.1 "gene75"    "gene60"    "gene65"    "gene66"    "gene1"    
#> cluster.2 "gene200"   "gene149"   "gene130"   "gene104"   "gene171"  
#> cluster.3 "gene232"   "gene297"   "gene289"   "gene275"   "gene234"  
#>           gene.top.28 gene.top.29 gene.top.30 gene.top.31 gene.top.32
#> cluster.1 "gene16"    "gene35"    "gene62"    "gene72"    "gene91"   
#> cluster.2 "gene155"   "gene184"   "gene121"   "gene120"   "gene114"  
#> cluster.3 "gene209"   "gene226"   "gene214"   "gene272"   "gene3752" 
#>           gene.top.33 gene.top.34 gene.top.35 gene.top.36 gene.top.37
#> cluster.1 "gene48"    "gene19"    "gene26"    "gene90"    "gene57"   
#> cluster.2 "gene119"   "gene153"   "gene160"   "gene141"   "gene166"  
#> cluster.3 "gene225"   "gene236"   "gene249"   "gene268"   "gene300"  
#>           gene.top.38 gene.top.39 gene.top.40 gene.top.41 gene.top.42
#> cluster.1 "gene79"    "gene70"    "gene47"    "gene5"     "gene82"   
#> cluster.2 "gene143"   "gene151"   "gene134"   "gene174"   "gene122"  
#> cluster.3 "gene252"   "gene208"   "gene276"   "gene285"   "gene295"  
#>           gene.top.43 gene.top.44 gene.top.45 gene.top.46 gene.top.47
#> cluster.1 "gene32"    "gene94"    "gene4"     "gene39"    "gene21"   
#> cluster.2 "gene112"   "gene157"   "gene113"   "gene128"   "gene193"  
#> cluster.3 "gene230"   "gene287"   "gene286"   "gene248"   "gene245"  
#>           gene.top.48 gene.top.49 gene.top.50 gene.top.51 gene.top.52
#> cluster.1 "gene63"    "gene50"    "gene88"    "gene86"    "gene43"   
#> cluster.2 "gene150"   "gene187"   "gene176"   "gene103"   "gene178"  
#> cluster.3 "gene207"   "gene299"   "gene205"   "gene269"   "gene283"  
#>           gene.top.53 gene.top.54 gene.top.55 gene.top.56 gene.top.57
#> cluster.1 "gene18"    "gene83"    "gene78"    "gene10"    "gene84"   
#> cluster.2 "gene139"   "gene189"   "gene101"   "gene127"   "gene106"  
#> cluster.3 "gene264"   "gene292"   "gene235"   "gene278"   "gene273"  
#>           gene.top.58 gene.top.59 gene.top.60 gene.top.61 gene.top.62
#> cluster.1 "gene53"    "gene85"    "gene7"     "gene100"   "gene24"   
#> cluster.2 "gene196"   "gene138"   "gene194"   "gene146"   "gene133"  
#> cluster.3 "gene298"   "gene256"   "gene231"   "gene2886"  "gene215"  
#>           gene.top.63 gene.top.64 gene.top.65 gene.top.66 gene.top.67
#> cluster.1 "gene80"    "gene9"     "gene96"    "gene46"    "gene76"   
#> cluster.2 "gene190"   "gene135"   "gene118"   "gene159"   "gene167"  
#> cluster.3 "gene203"   "gene270"   "gene251"   "gene288"   "gene243"  
#>           gene.top.68 gene.top.69 gene.top.70 gene.top.71 gene.top.72
#> cluster.1 "gene64"    "gene37"    "gene20"    "gene6"     "gene89"   
#> cluster.2 "gene163"   "gene170"   "gene126"   "gene144"   "gene172"  
#> cluster.3 "gene210"   "gene260"   "gene277"   "gene261"   "gene290"  
#>           gene.top.73 gene.top.74 gene.top.75 gene.top.76 gene.top.77
#> cluster.1 "gene55"    "gene71"    "gene30"    "gene67"    "gene23"   
#> cluster.2 "gene136"   "gene186"   "gene142"   "gene185"   "gene181"  
#> cluster.3 "gene266"   "gene798"   "gene206"   "gene259"   "gene281"  
#>           gene.top.78 gene.top.79 gene.top.80 gene.top.81 gene.top.82
#> cluster.1 "gene97"    "gene41"    "gene40"    "gene42"    "gene69"   
#> cluster.2 "gene137"   "gene156"   "gene117"   "gene197"   "gene158"  
#> cluster.3 "gene254"   "gene293"   "gene2694"  "gene239"   "gene213"  
#>           gene.top.83 gene.top.84 gene.top.85 gene.top.86 gene.top.87
#> cluster.1 "gene61"    "gene95"    "gene11"    "gene2"     "gene13"   
#> cluster.2 "gene169"   "gene175"   "gene105"   "gene129"   "gene165"  
#> cluster.3 "gene201"   "gene233"   NA          NA          NA         
#>           gene.top.88 gene.top.89 gene.top.90 gene.top.91 gene.top.92
#> cluster.1 "gene44"    "gene17"    "gene14"    "gene73"    "gene22"   
#> cluster.2 "gene195"   "gene132"   "gene110"   "gene182"   "gene162"  
#> cluster.3 NA          NA          NA          NA          NA         
#>           gene.top.93 gene.top.94 gene.top.95 gene.top.96 gene.top.97
#> cluster.1 "gene38"    "gene81"    "gene56"    "gene34"    "gene45"   
#> cluster.2 "gene111"   "gene124"   "gene131"   "gene145"   "gene109"  
#> cluster.3 NA          NA          NA          NA          NA         
#>           gene.top.98 gene.top.99 gene.top.100 gene.top.101 gene.top.102
#> cluster.1 "gene99"    "gene52"    "gene93"     "gene916"    "gene1009"  
#> cluster.2 "gene147"   "gene188"   "gene116"    "gene3172"   "gene3443"  
#> cluster.3 NA          NA          NA           NA           NA          
#>           gene.top.103 gene.top.104 gene.top.105 gene.top.106 gene.top.107
#> cluster.1 "gene3213"   "gene2326"   "gene2356"   "gene1338"   "gene3956"  
#> cluster.2 "gene1672"   "gene618"    "gene2616"   "gene1765"   "gene3606"  
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.108 gene.top.109 gene.top.110 gene.top.111 gene.top.112
#> cluster.1 "gene946"    "gene2312"   "gene3231"   "gene2242"   NA          
#> cluster.2 "gene3812"   "gene632"    "gene2182"   "gene1356"   NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.113 gene.top.114 gene.top.115 gene.top.116 gene.top.117
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.118 gene.top.119 gene.top.120 gene.top.121 gene.top.122
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.123 gene.top.124 gene.top.125 gene.top.126 gene.top.127
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.128 gene.top.129 gene.top.130 gene.top.131 gene.top.132
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.133 gene.top.134 gene.top.135 gene.top.136 gene.top.137
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.138 gene.top.139 gene.top.140 gene.top.141 gene.top.142
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.143 gene.top.144 gene.top.145 gene.top.146 gene.top.147
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.148 gene.top.149 gene.top.150 gene.top.151 gene.top.152
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.153 gene.top.154 gene.top.155 gene.top.156 gene.top.157
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.158 gene.top.159 gene.top.160 gene.top.161 gene.top.162
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.163 gene.top.164 gene.top.165 gene.top.166 gene.top.167
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.168 gene.top.169 gene.top.170 gene.top.171 gene.top.172
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.173 gene.top.174 gene.top.175 gene.top.176 gene.top.177
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.178 gene.top.179 gene.top.180 gene.top.181 gene.top.182
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.183 gene.top.184 gene.top.185 gene.top.186 gene.top.187
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.188 gene.top.189 gene.top.190 gene.top.191 gene.top.192
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.193 gene.top.194 gene.top.195 gene.top.196 gene.top.197
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.198 gene.top.199 gene.top.200 gene.top.201 gene.top.202
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.203 gene.top.204 gene.top.205 gene.top.206 gene.top.207
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.208 gene.top.209 gene.top.210 gene.top.211 gene.top.212
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.213 gene.top.214 gene.top.215 gene.top.216 gene.top.217
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.218 gene.top.219 gene.top.220 gene.top.221 gene.top.222
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.223 gene.top.224 gene.top.225 gene.top.226 gene.top.227
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.228 gene.top.229 gene.top.230 gene.top.231 gene.top.232
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.233 gene.top.234 gene.top.235 gene.top.236 gene.top.237
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.238 gene.top.239 gene.top.240 gene.top.241 gene.top.242
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.243 gene.top.244 gene.top.245 gene.top.246 gene.top.247
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.248 gene.top.249 gene.top.250 gene.top.251 gene.top.252
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.253 gene.top.254 gene.top.255 gene.top.256 gene.top.257
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.258 gene.top.259 gene.top.260 gene.top.261 gene.top.262
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.263 gene.top.264 gene.top.265 gene.top.266 gene.top.267
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.268 gene.top.269 gene.top.270 gene.top.271 gene.top.272
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.273 gene.top.274 gene.top.275 gene.top.276 gene.top.277
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.278 gene.top.279 gene.top.280 gene.top.281 gene.top.282
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.283 gene.top.284 gene.top.285 gene.top.286 gene.top.287
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.288 gene.top.289 gene.top.290 gene.top.291 gene.top.292
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.293 gene.top.294 gene.top.295 gene.top.296 gene.top.297
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.298 gene.top.299 gene.top.300 gene.top.301 gene.top.302
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.303 gene.top.304 gene.top.305 gene.top.306 gene.top.307
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.308 gene.top.309 gene.top.310 gene.top.311 gene.top.312
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.313 gene.top.314 gene.top.315 gene.top.316 gene.top.317
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.318 gene.top.319 gene.top.320 gene.top.321 gene.top.322
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.323 gene.top.324 gene.top.325 gene.top.326 gene.top.327
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.328 gene.top.329 gene.top.330 gene.top.331 gene.top.332
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.333 gene.top.334 gene.top.335 gene.top.336 gene.top.337
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.338 gene.top.339 gene.top.340 gene.top.341 gene.top.342
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.343 gene.top.344 gene.top.345 gene.top.346 gene.top.347
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.348 gene.top.349 gene.top.350 gene.top.351 gene.top.352
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.353 gene.top.354 gene.top.355 gene.top.356 gene.top.357
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.358 gene.top.359 gene.top.360 gene.top.361 gene.top.362
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.363 gene.top.364 gene.top.365 gene.top.366 gene.top.367
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.368 gene.top.369 gene.top.370 gene.top.371 gene.top.372
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.373 gene.top.374 gene.top.375 gene.top.376 gene.top.377
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.378 gene.top.379 gene.top.380 gene.top.381 gene.top.382
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.383 gene.top.384 gene.top.385 gene.top.386 gene.top.387
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.388 gene.top.389 gene.top.390 gene.top.391 gene.top.392
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.393 gene.top.394 gene.top.395 gene.top.396 gene.top.397
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.398 gene.top.399 gene.top.400 gene.top.401 gene.top.402
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.403 gene.top.404 gene.top.405 gene.top.406 gene.top.407
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.408 gene.top.409 gene.top.410 gene.top.411 gene.top.412
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.413 gene.top.414 gene.top.415 gene.top.416 gene.top.417
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.418 gene.top.419 gene.top.420 gene.top.421 gene.top.422
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.423 gene.top.424 gene.top.425 gene.top.426 gene.top.427
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.428 gene.top.429 gene.top.430 gene.top.431 gene.top.432
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.433 gene.top.434 gene.top.435 gene.top.436 gene.top.437
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.438 gene.top.439 gene.top.440 gene.top.441 gene.top.442
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.443 gene.top.444 gene.top.445 gene.top.446 gene.top.447
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.448 gene.top.449 gene.top.450 gene.top.451 gene.top.452
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.453 gene.top.454 gene.top.455 gene.top.456 gene.top.457
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.458 gene.top.459 gene.top.460 gene.top.461 gene.top.462
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.463 gene.top.464 gene.top.465 gene.top.466 gene.top.467
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.468 gene.top.469 gene.top.470 gene.top.471 gene.top.472
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.473 gene.top.474 gene.top.475 gene.top.476 gene.top.477
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.478 gene.top.479 gene.top.480 gene.top.481 gene.top.482
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.483 gene.top.484 gene.top.485 gene.top.486 gene.top.487
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.488 gene.top.489 gene.top.490 gene.top.491 gene.top.492
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.493 gene.top.494 gene.top.495 gene.top.496 gene.top.497
#> cluster.1 NA           NA           NA           NA           NA          
#> cluster.2 NA           NA           NA           NA           NA          
#> cluster.3 NA           NA           NA           NA           NA          
#>           gene.top.498 gene.top.499 gene.top.500
#> cluster.1 NA           NA           NA          
#> cluster.2 NA           NA           NA          
#> cluster.3 NA           NA           NA          

if (FALSE) {
## with an artificial dataset

m <- matrix(rnorm(80000), nc = 20)
res <- get_data_4_DBFMCL(data=m)
}