Transform a matrix

This function applies a series of transformations to a matrix, including log2 transformation, centering, standardizing, ceiling, and flooring.

transform_mat(
  m,
  to_log2 = FALSE,
  pseudocount = 1,
  centering = TRUE,
  standardizing = FALSE,
  ceil = NULL,
  floor = NULL
)

Arguments

m

A numeric matrix to be transformed.

to_log2

A logical indicating whether to apply log2 transformation (default is FALSE).

pseudocount

A numeric value to be added to each element before log2 transformation (default is 1).

centering

A logical indicating whether to center the rows by subtracting the mean (default is TRUE).

standardizing

A logical indicating whether to standardize the rows by dividing by the standard deviation (default is FALSE).

ceil

A numeric value for ceiling (default is NULL, meaning no ceiling is applied).

floor

A numeric value for flooring (default is NULL, meaning no flooring is applied).

Value

The transformed matrix.

Examples

#' # Create a sample matrix
mat <- matrix(1:12, nrow=3)
#' # Transform the matrix
transformed_mat <- transform_mat(mat, to_log2=TRUE, centering=TRUE, standardizing=TRUE, ceil=2, floor=-2)