Best co-expressed genes from each gene cluster

Extract the highly co-expressed genes of each gene cluster.

top_genes(object, top = 20, cluster = "all")

Arguments

object

A ClusterSet object.

top

A value for the number of the most similar genes in gene clusters.

cluster

A vector of gene cluster identity.

Value

Examples

m <- matrix(rnorm(80000), nc=20)
m[1:100,1:10] <- m[1:100,1:10] + 4
m[101:200,11:20] <- m[101:200,11:20] + 3
m[201:300,5:15] <- m[201:300,5:15] + -2

res <- DBFMCL(data=m,
              distance_method="pearson",
              av_dot_prod_min = 0,
              inflation = 1.2,
              k=25,
              fdr = 10)
#> |--  Row names not provided. Adding. 
#> |--  Colum names not provided. Adding. 
#> The following parameters will be used : 
#> 	Working directory:  /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference 
#> 	Ouput directory:  /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference 
#> 	Name:  X1tR0oi30C 
#> 	Distance method:  pearson 
#> 	Minimum average dot product for clusters:  0 
#> 	Minimum cluster size:  10 
#> 	Number of neighbors:  25 
#> 	FDR:  10 % 
#> 	Inflation: 1.2 
#> 	Visualize standard outputs from both mcl and cluster commands:  FALSE 
#> 
#> |--  DBF completed. Starting MCL step. 
#> Running mcl (graph partitioning)... 
#> |--  Done 
#> |--  creating file : /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference/X1tR0oi30C.mcl_out.txt 
#> |--  Reading and filtering MCL output: /mnt/NAS7/Workspace/bavaisj/dbfmcl/scigenex/docs/reference/X1tR0oi30C.mcl_out.txt 
#> |--  3  clusters conserved after MCL partitioning. 
#> |--  0  clusters filtered out from MCL partitioning (size and mean dot product). 
              
res <- top_genes(object = res, top=500, cluster="all")
#> Warning: Number of top genes is greater than the number of genes in cluster 1. All genes will be use and order by similarity rank.
#> Warning: Number of top genes is greater than the number of genes in cluster 2. All genes will be use and order by similarity rank.
#> Warning: Number of top genes is greater than the number of genes in cluster 3. All genes will be use and order by similarity rank.
#> |--  Results are stored in top_genes slot of ClusterSet object. 
res@top_genes
#>           gene_top_1 gene_top_2 gene_top_3 gene_top_4 gene_top_5 gene_top_6
#> cluster_1 "gene68"   "gene58"   "gene92"   "gene8"    "gene27"   "gene36"  
#> cluster_2 "gene164"  "gene152"  "gene192"  "gene180"  "gene199"  "gene125" 
#> cluster_3 "gene250"  "gene227"  "gene257"  "gene212"  "gene253"  "gene221" 
#>           gene_top_7 gene_top_8 gene_top_9 gene_top_10 gene_top_11 gene_top_12
#> cluster_1 "gene49"   "gene54"   "gene3"    "gene59"    "gene77"    "gene33"   
#> cluster_2 "gene161"  "gene148"  "gene177"  "gene115"   "gene102"   "gene107"  
#> cluster_3 "gene217"  "gene271"  "gene244"  "gene246"   "gene258"   "gene216"  
#>           gene_top_13 gene_top_14 gene_top_15 gene_top_16 gene_top_17
#> cluster_1 "gene98"    "gene31"    "gene25"    "gene51"    "gene29"   
#> cluster_2 "gene173"   "gene183"   "gene123"   "gene108"   "gene198"  
#> cluster_3 "gene211"   "gene228"   "gene263"   "gene229"   "gene240"  
#>           gene_top_18 gene_top_19 gene_top_20 gene_top_21 gene_top_22
#> cluster_1 "gene15"    "gene87"    "gene74"    "gene28"    "gene12"   
#> cluster_2 "gene154"   "gene179"   "gene191"   "gene140"   "gene168"  
#> cluster_3 "gene267"   "gene265"   "gene279"   "gene204"   "gene280"  
#>           gene_top_23 gene_top_24 gene_top_25 gene_top_26 gene_top_27
#> cluster_1 "gene75"    "gene60"    "gene65"    "gene66"    "gene1"    
#> cluster_2 "gene200"   "gene149"   "gene130"   "gene104"   "gene171"  
#> cluster_3 "gene232"   "gene297"   "gene289"   "gene275"   "gene234"  
#>           gene_top_28 gene_top_29 gene_top_30 gene_top_31 gene_top_32
#> cluster_1 "gene16"    "gene35"    "gene62"    "gene72"    "gene91"   
#> cluster_2 "gene155"   "gene184"   "gene121"   "gene120"   "gene114"  
#> cluster_3 "gene209"   "gene226"   "gene214"   "gene272"   "gene3752" 
#>           gene_top_33 gene_top_34 gene_top_35 gene_top_36 gene_top_37
#> cluster_1 "gene48"    "gene19"    "gene26"    "gene90"    "gene57"   
#> cluster_2 "gene119"   "gene153"   "gene160"   "gene141"   "gene166"  
#> cluster_3 "gene225"   "gene236"   "gene249"   "gene268"   "gene300"  
#>           gene_top_38 gene_top_39 gene_top_40 gene_top_41 gene_top_42
#> cluster_1 "gene79"    "gene70"    "gene47"    "gene5"     "gene82"   
#> cluster_2 "gene143"   "gene151"   "gene134"   "gene174"   "gene122"  
#> cluster_3 "gene252"   "gene208"   "gene276"   "gene285"   "gene295"  
#>           gene_top_43 gene_top_44 gene_top_45 gene_top_46 gene_top_47
#> cluster_1 "gene32"    "gene94"    "gene4"     "gene39"    "gene21"   
#> cluster_2 "gene112"   "gene157"   "gene113"   "gene128"   "gene193"  
#> cluster_3 "gene230"   "gene287"   "gene286"   "gene248"   "gene245"  
#>           gene_top_48 gene_top_49 gene_top_50 gene_top_51 gene_top_52
#> cluster_1 "gene63"    "gene50"    "gene88"    "gene86"    "gene43"   
#> cluster_2 "gene150"   "gene187"   "gene176"   "gene103"   "gene178"  
#> cluster_3 "gene207"   "gene299"   "gene205"   "gene269"   "gene283"  
#>           gene_top_53 gene_top_54 gene_top_55 gene_top_56 gene_top_57
#> cluster_1 "gene18"    "gene83"    "gene78"    "gene10"    "gene84"   
#> cluster_2 "gene139"   "gene189"   "gene101"   "gene127"   "gene106"  
#> cluster_3 "gene264"   "gene292"   "gene235"   "gene278"   "gene273"  
#>           gene_top_58 gene_top_59 gene_top_60 gene_top_61 gene_top_62
#> cluster_1 "gene53"    "gene85"    "gene7"     "gene100"   "gene24"   
#> cluster_2 "gene196"   "gene138"   "gene194"   "gene146"   "gene133"  
#> cluster_3 "gene298"   "gene256"   "gene231"   "gene2886"  "gene215"  
#>           gene_top_63 gene_top_64 gene_top_65 gene_top_66 gene_top_67
#> cluster_1 "gene80"    "gene9"     "gene96"    "gene46"    "gene76"   
#> cluster_2 "gene190"   "gene135"   "gene118"   "gene159"   "gene167"  
#> cluster_3 "gene203"   "gene270"   "gene251"   "gene288"   "gene243"  
#>           gene_top_68 gene_top_69 gene_top_70 gene_top_71 gene_top_72
#> cluster_1 "gene64"    "gene37"    "gene20"    "gene6"     "gene89"   
#> cluster_2 "gene163"   "gene170"   "gene126"   "gene144"   "gene172"  
#> cluster_3 "gene210"   "gene260"   "gene277"   "gene261"   "gene290"  
#>           gene_top_73 gene_top_74 gene_top_75 gene_top_76 gene_top_77
#> cluster_1 "gene55"    "gene71"    "gene30"    "gene67"    "gene23"   
#> cluster_2 "gene136"   "gene186"   "gene142"   "gene185"   "gene181"  
#> cluster_3 "gene266"   "gene798"   "gene206"   "gene259"   "gene281"  
#>           gene_top_78 gene_top_79 gene_top_80 gene_top_81 gene_top_82
#> cluster_1 "gene97"    "gene41"    "gene40"    "gene42"    "gene69"   
#> cluster_2 "gene137"   "gene156"   "gene117"   "gene197"   "gene158"  
#> cluster_3 "gene254"   "gene293"   "gene2694"  "gene239"   "gene213"  
#>           gene_top_83 gene_top_84 gene_top_85 gene_top_86 gene_top_87
#> cluster_1 "gene61"    "gene95"    "gene11"    "gene2"     "gene13"   
#> cluster_2 "gene169"   "gene175"   "gene105"   "gene129"   "gene165"  
#> cluster_3 "gene201"   "gene233"   NA          NA          NA         
#>           gene_top_88 gene_top_89 gene_top_90 gene_top_91 gene_top_92
#> cluster_1 "gene44"    "gene17"    "gene14"    "gene73"    "gene22"   
#> cluster_2 "gene195"   "gene132"   "gene110"   "gene182"   "gene162"  
#> cluster_3 NA          NA          NA          NA          NA         
#>           gene_top_93 gene_top_94 gene_top_95 gene_top_96 gene_top_97
#> cluster_1 "gene38"    "gene81"    "gene56"    "gene34"    "gene45"   
#> cluster_2 "gene111"   "gene124"   "gene131"   "gene145"   "gene109"  
#> cluster_3 NA          NA          NA          NA          NA         
#>           gene_top_98 gene_top_99 gene_top_100 gene_top_101 gene_top_102
#> cluster_1 "gene99"    "gene52"    "gene93"     "gene916"    "gene1009"  
#> cluster_2 "gene147"   "gene188"   "gene116"    "gene3172"   "gene3443"  
#> cluster_3 NA          NA          NA           NA           NA          
#>           gene_top_103 gene_top_104 gene_top_105 gene_top_106 gene_top_107
#> cluster_1 "gene3213"   "gene2326"   "gene2356"   "gene1338"   "gene3956"  
#> cluster_2 "gene1672"   "gene618"    "gene2616"   "gene1765"   "gene3606"  
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_108 gene_top_109 gene_top_110 gene_top_111 gene_top_112
#> cluster_1 "gene946"    "gene2312"   "gene3231"   "gene2242"   NA          
#> cluster_2 "gene3812"   "gene632"    "gene2182"   "gene1356"   NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_113 gene_top_114 gene_top_115 gene_top_116 gene_top_117
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_118 gene_top_119 gene_top_120 gene_top_121 gene_top_122
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_123 gene_top_124 gene_top_125 gene_top_126 gene_top_127
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_128 gene_top_129 gene_top_130 gene_top_131 gene_top_132
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_133 gene_top_134 gene_top_135 gene_top_136 gene_top_137
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_138 gene_top_139 gene_top_140 gene_top_141 gene_top_142
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_143 gene_top_144 gene_top_145 gene_top_146 gene_top_147
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_148 gene_top_149 gene_top_150 gene_top_151 gene_top_152
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_153 gene_top_154 gene_top_155 gene_top_156 gene_top_157
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_158 gene_top_159 gene_top_160 gene_top_161 gene_top_162
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_163 gene_top_164 gene_top_165 gene_top_166 gene_top_167
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_168 gene_top_169 gene_top_170 gene_top_171 gene_top_172
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_173 gene_top_174 gene_top_175 gene_top_176 gene_top_177
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_178 gene_top_179 gene_top_180 gene_top_181 gene_top_182
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_183 gene_top_184 gene_top_185 gene_top_186 gene_top_187
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_188 gene_top_189 gene_top_190 gene_top_191 gene_top_192
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_193 gene_top_194 gene_top_195 gene_top_196 gene_top_197
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_198 gene_top_199 gene_top_200 gene_top_201 gene_top_202
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_203 gene_top_204 gene_top_205 gene_top_206 gene_top_207
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_208 gene_top_209 gene_top_210 gene_top_211 gene_top_212
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_213 gene_top_214 gene_top_215 gene_top_216 gene_top_217
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_218 gene_top_219 gene_top_220 gene_top_221 gene_top_222
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_223 gene_top_224 gene_top_225 gene_top_226 gene_top_227
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_228 gene_top_229 gene_top_230 gene_top_231 gene_top_232
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_233 gene_top_234 gene_top_235 gene_top_236 gene_top_237
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_238 gene_top_239 gene_top_240 gene_top_241 gene_top_242
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_243 gene_top_244 gene_top_245 gene_top_246 gene_top_247
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_248 gene_top_249 gene_top_250 gene_top_251 gene_top_252
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_253 gene_top_254 gene_top_255 gene_top_256 gene_top_257
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_258 gene_top_259 gene_top_260 gene_top_261 gene_top_262
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_263 gene_top_264 gene_top_265 gene_top_266 gene_top_267
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_268 gene_top_269 gene_top_270 gene_top_271 gene_top_272
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_273 gene_top_274 gene_top_275 gene_top_276 gene_top_277
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_278 gene_top_279 gene_top_280 gene_top_281 gene_top_282
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_283 gene_top_284 gene_top_285 gene_top_286 gene_top_287
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_288 gene_top_289 gene_top_290 gene_top_291 gene_top_292
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_293 gene_top_294 gene_top_295 gene_top_296 gene_top_297
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_298 gene_top_299 gene_top_300 gene_top_301 gene_top_302
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_303 gene_top_304 gene_top_305 gene_top_306 gene_top_307
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_308 gene_top_309 gene_top_310 gene_top_311 gene_top_312
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_313 gene_top_314 gene_top_315 gene_top_316 gene_top_317
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_318 gene_top_319 gene_top_320 gene_top_321 gene_top_322
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_323 gene_top_324 gene_top_325 gene_top_326 gene_top_327
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_328 gene_top_329 gene_top_330 gene_top_331 gene_top_332
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_333 gene_top_334 gene_top_335 gene_top_336 gene_top_337
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_338 gene_top_339 gene_top_340 gene_top_341 gene_top_342
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_343 gene_top_344 gene_top_345 gene_top_346 gene_top_347
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_348 gene_top_349 gene_top_350 gene_top_351 gene_top_352
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_353 gene_top_354 gene_top_355 gene_top_356 gene_top_357
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_358 gene_top_359 gene_top_360 gene_top_361 gene_top_362
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_363 gene_top_364 gene_top_365 gene_top_366 gene_top_367
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_368 gene_top_369 gene_top_370 gene_top_371 gene_top_372
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_373 gene_top_374 gene_top_375 gene_top_376 gene_top_377
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_378 gene_top_379 gene_top_380 gene_top_381 gene_top_382
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_383 gene_top_384 gene_top_385 gene_top_386 gene_top_387
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_388 gene_top_389 gene_top_390 gene_top_391 gene_top_392
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_393 gene_top_394 gene_top_395 gene_top_396 gene_top_397
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_398 gene_top_399 gene_top_400 gene_top_401 gene_top_402
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_403 gene_top_404 gene_top_405 gene_top_406 gene_top_407
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_408 gene_top_409 gene_top_410 gene_top_411 gene_top_412
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_413 gene_top_414 gene_top_415 gene_top_416 gene_top_417
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_418 gene_top_419 gene_top_420 gene_top_421 gene_top_422
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_423 gene_top_424 gene_top_425 gene_top_426 gene_top_427
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_428 gene_top_429 gene_top_430 gene_top_431 gene_top_432
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_433 gene_top_434 gene_top_435 gene_top_436 gene_top_437
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_438 gene_top_439 gene_top_440 gene_top_441 gene_top_442
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_443 gene_top_444 gene_top_445 gene_top_446 gene_top_447
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_448 gene_top_449 gene_top_450 gene_top_451 gene_top_452
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_453 gene_top_454 gene_top_455 gene_top_456 gene_top_457
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_458 gene_top_459 gene_top_460 gene_top_461 gene_top_462
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_463 gene_top_464 gene_top_465 gene_top_466 gene_top_467
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_468 gene_top_469 gene_top_470 gene_top_471 gene_top_472
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_473 gene_top_474 gene_top_475 gene_top_476 gene_top_477
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_478 gene_top_479 gene_top_480 gene_top_481 gene_top_482
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_483 gene_top_484 gene_top_485 gene_top_486 gene_top_487
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_488 gene_top_489 gene_top_490 gene_top_491 gene_top_492
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_493 gene_top_494 gene_top_495 gene_top_496 gene_top_497
#> cluster_1 NA           NA           NA           NA           NA          
#> cluster_2 NA           NA           NA           NA           NA          
#> cluster_3 NA           NA           NA           NA           NA          
#>           gene_top_498 gene_top_499 gene_top_500
#> cluster_1 NA           NA           NA          
#> cluster_2 NA           NA           NA          
#> cluster_3 NA           NA           NA          

if (FALSE) {
## with an artificial dataset

m <- matrix(rnorm(80000), nc = 20)
res <- get_data_4_DBFMCL(data=m)
}